requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.3
django-admin==2.0.2
django-allauth==0.61.1
django-cors-headers==4.3.1
django-excel-response2==3.0.6
django-filter==24.2
django-six==1.0.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
excel-base==1.0.4
idna==3.6
inflection==0.5.1
isoweek==1.3.3
Markdown==3.6
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
pretty-errors==1.2.25
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
xlwt==1.3.0
